requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
numpy==1.19.4
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.4
six==1.15.0
tornado==6.1
tzlocal==2.1
